use assert system with virtual function utilities and have better names
