+++ implementation/virtual-function-utilities/stack-based_virtual.h
@@ -3,7 +3,7 @@
 #include<array>
 #include<string>
 #include<typeinfo>
-#include<assert.h>
+#include "../asserts_and_errors.h"
 #include<functional>
 
 namespace mu
@@ -225,7 +225,7 @@ public:
     void downcast(stack_virt<d,cap>& ret, std::function<void(base*,base*)> const& move_functor) &&
     {
         static_assert((std::is_base_of<base,d>::value || std::is_same<base,d>::value) && sizeof(d) <= cap,"to downcast stack_virt<xb,xc>&& to y*, yb must derive from xb and");
-        assert(can_downcast<d>());
+		expr::impl::assert_with_generic_logic_error(can_downcast<d>());
         if(ret.check_state_whether_nonnull())
         {
             ret.call_destructor_on_data();
@@ -243,7 +243,7 @@ public:
     d* downcast_get()
     {
         static_assert((std::is_base_of<base,d>::value || std::is_same<base,d>::value) && sizeof(d) <= cap,"to downcast stack_virt<xb,xc>&& to y*, yb must derive from xb and");
-        assert(can_downcast<d>());
+		expr::impl::assert_with_generic_logic_error(can_downcast<d>());
         return static_cast<d*>(get());
     }
 
@@ -251,7 +251,7 @@ public:
     d const* downcast_get() const
     {
         static_assert((std::is_base_of<base,d>::value || std::is_same<base,d>::value) && sizeof(d) <= cap,"to downcast stack_virt<xb,xc>&& to y*, yb must derive from xb and");
-        assert(can_downcast<d>());
+		expr::impl::assert_with_generic_logic_error(can_downcast<d>());
         return static_cast<d const*>(get());
     }
 
@@ -267,7 +267,7 @@ public:
 
     stack_virt<base,sizeof(base)> shrink(stack_virt<base,sizeof(base)>& ret, std::function<void(base*,base*)> const& move_functor) &&
     {
-        assert(is_only_base());
+		expr::impl::assert_with_generic_logic_error(is_only_base());
 
         if(ret.check_state_whether_nonnull())
         {
